srcs/API-Gateway/requirements.txt
Django==4.2
djangorestframework
django-otp
python-dotenv
psycopg2
gunicorn
requests